+++ src/Pages/Proposal/Components/ProposalDepositsChart.tsx
@@ -9,7 +9,7 @@ import Big from 'big.js';
 const StyledChart = styled.div`
   height: 200px;
   width: 100%;
-  margin-top: -60px;
+  margin-top: -45px;
   margin-bottom: -70px;
 `;
 
@@ -67,6 +67,31 @@ const chartData = {
         shadowOffsetY: 1,
         borderRadius: [5,5,5,5],
       },
+      markLine: {
+        silent: true,
+        data: [
+          {
+            name: 'Deposit Threshold',
+            xAxis: -1,
+            label: {
+              show: true,
+              formatter: 'Pass Threshold (50,000 hash)',
+              fontStyle: 'normal',
+              fontFamily: 'Montserrat',
+              textBorderColor: 'none',
+              color: '',
+            },
+            lineStyle: {
+              color: 'red',
+            },
+          },
+        ],
+        symbol: 'line',
+        lineStyle: {
+          color: 'black',
+          width: 2,
+        },
+      },
     },
   ]
 };
@@ -89,7 +114,7 @@ export const ProposalDepositsChart = () => {
     () => {
       // Build data
       const deposits = {
-        value: getPercentage(current, needed) > 100 ? 100 : getPercentage(current, needed),
+        value: getPercentage(current, needed),
         initialDeposit: initial,
         currentDeposit: current,
       };
@@ -110,6 +135,14 @@ export const ProposalDepositsChart = () => {
       chartData.series[0].data = [deposits];
       chartData.series[0].backgroundStyle.color = theme.BACKGROUND_LIGHT;
       chartData.series[0].backgroundStyle.shadowColor = theme.BACKGROUND_BLACK;
+      if (getPercentage(current, needed) >= 100) {
+        chartData.series[0].markLine.data[0].xAxis = getPercentage(needed, current);
+        chartData.series[0].markLine.data[0].label.color = theme.FONT_PRIMARY;
+      }
+      else {
+        chartData.series[0].markLine.data[0].xAxis = 100;
+        chartData.series[0].markLine.data[0].label.color = theme.FONT_PRIMARY;
+      }
 
       chartData.tooltip.formatter = (params: ParamsArray[]) => (
         `<div style="padding:2px;">