+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Features
 - FE updates to proposal view #332
+- Add markline to proposal deposits chart #366
 
 ### Features